balanced criteria to avoid explosive reproduction
fixed the bug when both parents were giving birth to one kid each. Added the  simulation ending message.

+++ initialparameters.h
@@ -18,7 +18,7 @@ public:
     static int predsHp;
     static double herbsRepSatiety;
     static double predsRepSatiety;
-    static int omsRepSatiety;
+    static double omsRepSatiety;
     static int minHerbsLifetime;
     static int minPredsLifetime;
     static int minOmsLifetime;
